requirements.txt
# Deep Learning
tensorflow>=2.12.0

# Procesamiento de datos
numpy>=1.24.0
pandas>=2.0.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.2.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.0.0

# Utilidades
tqdm>=4.65.0
Pillow>=9.5.0




# RAG específico
langchain>=0.3.0
langchain-community>=0.3.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.25.0
openai>=1.0.0  # o usar ollama
ollama>=0.1.0

